feat: Add selectGroup selector
