+++ frontend/components/group/group_show_container.jsx
@@ -1,20 +1,10 @@
 import { connect } from "react-redux";
 import { fetchGroup } from "../../actions/group_actions";
 import GroupShow from "./group_show";
-import { selectCurrentUser } from "../../reducers/selectors";
-
-const DEFAULT_GROUP = {
-  name: "",
-  description: "",
-  group_user_ids: [],
-  hometown: "",
-  id: null,
-  member_ids: [],
-};
+import { selectGroup } from "../../reducers/selectors";
 
 const msp = (state, ownProps) => ({
-  group: state.entities.groups[ownProps.match.params.groupId] || DEFAULT_GROUP,
-  currentUser: selectCurrentUser(state),
+  group: selectGroup(state, ownProps.match.params.groupId),
 });
 
 const mdp = dispatch => ({
