+++ frontend/reducers/selectors.js
@@ -1,3 +1,30 @@
+import merge from "lodash/merge";
+
+export const selectGroup = (state, groupId) => {
+  const DEFAULT_GROUP = {
+    name: "",
+    description: "",
+    group_user_ids: [],
+    hometown: "",
+    id: null,
+    member_ids: [],
+  };
+  const originalGroup = state.entities.groups[groupId] || {};
+  const newGroup = merge({}, DEFAULT_GROUP, originalGroup);
+  newGroup.organizerIds = [];
+  newGroup.group_user_ids.forEach(groupUserId => {
+    const groupUser = state.entities.groupUsers[groupUserId];
+    const isOrganizer = groupUser.member_type === 'Organizer';
+    if (isOrganizer){
+      newGroup.organizerIds.push(groupUser.user_id);
+    }
+    if (state.session.id === groupUser.user_id) newGroup.isMember = true;
+    console.log('newGroup.isMember: ', newGroup.isMember);
+    if (newGroup.isMember && isOrganizer) newGroup.isOrganizer = true;
+  });
+  return newGroup;
+};
+
 export const selectCurrentUser = ({ entities, session }) => {
   return entities.users[session.id];
 };
